// Ethix/ethix-ng-demo/src/models/CustomerAccount.ts
class CustomerAccount {
    accId: number;
    Acc_Type: string;
    Acc_Number: string;
    Class_Code: string;
    openning_Bal: number;
    IsClosed: Boolean;
    currencyId: number;
    CustomerId: number;
    currency: number;
    currDecimal:number;
}

export default CustomerAccount;